% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_models.R
\name{prl_sim_traits_continuous}
\alias{prl_sim_traits_continuous}
\title{Title}
\usage{
prl_sim_traits_continuous(x, family = c("gaussian"), n_trait = 1,
  root_value = 0, initial_rate, tau = 1, sd = 0.25)
}
\arguments{
\item{x}{Object to be simulated from. Can be either a \code{phurl} object generated from
\code{\link{phurl_prepare_data}}, or an object of class \code{phylo}}

\item{family}{What error structure to use in the simulations? Can be a character vector of length \code{n_trait}
to use different families for different traits.}

\item{n_trait}{Number of traits to simulate.}

\item{root_value}{A vector of root values for each trait. Will be recycled to length \code{n_trait}}

\item{initial_rate}{A vector of initial rates. Rate at which traits are diverging after first split (will
be assumed to be c(initial_rate, -initial_rate) along the first splitting edges). Will be recycled to length 
\code{n_trait}}

\item{tau}{A vector tau values. This parameter controls the degree to which trait evolution rates change at
phylogeny nodes. Rate changes are sampled from a gaussian distribution with mean zero and sd equal to tau.
Will be recycled to length \code{n_trait}}

\item{sd}{Gaussian error of final tip-level observations. Standard deviation of noise added to tip-level
trait values. Only used if \code{family="gaussian"}}
}
\description{
Title
}
