% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phurl_prepare_data.R
\name{phurl_prepare_data}
\alias{phurl_prepare_data}
\title{Prepare data for analysing with phurl models}
\usage{
phurl_prepare_data(tree, traits, species_column = NULL,
  scale_edges = TRUE, centre = TRUE, transform_traits = NULL,
  standardise_traits = TRUE)
}
\arguments{
\item{tree}{Phylogenetic tree of class \code{\link[ape]{phylo}} or \code{\link[tibble:tbl_df-class]{tibble}}
(e.g. using \code{\link[tidytree]{as_tibble}})}

\item{traits}{Trait data. Can be a named \code{vector}, a \code{matrix} or \code{data.frame} with named rows, where the rownames
are the species names, or a \code{data.frame} or \code{\link[tibble:tbl_df-class]{tibble}} with a column specifying the
species name. If species names are not referenced in either the rownames or a column, data will be matched
to the tree using order only (e.g. assuming data rows are in the same order as \code{tree$tip.label}).}

\item{species_column}{Character value specifying which column of \code{traits} refers to the species
names, if they are not specified in the rownames. If not \code{NULL}, this will override any rownames in
\code{traits}}

\item{scale_edges}{Should the edges be scaled to a maximum of 1? This is highly recommended for regularised
models. Default TRUE.}

\item{centre}{Should the resulting edge path predictor matrix be centred to so that each column has a mean
of 0? This is recommended to improve convergence of the models. Default TRUE.}

\item{transform_traits}{List of transformations to apply to trait values. If named, will be matched to column
names, else they will be applied in order, and recycled if necessary.}

\item{standardise_traits}{Should traits values be standardised by subtracting the mean and dividing by
them standard deviation? Recommended. Default TRUE.}
}
\value{
A object of class \code{\link{phurl}}
}
\description{
Prepare data for analysing with phurl models
}
